# swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Various APIs for interacting with authentication, users, and tokens."
  version: "v1"
  title: "Stroom Auth API"
  contact:
    name: "GCHQ"
    url: "https://github.com/gchq/stroom-auth"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /user/v1:
    get:
      tags:
      - "User"
      summary: "Get all users."
      description: ""
      operationId: "getAll"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
    post:
      tags:
      - "User"
      summary: "Create a user."
      description: ""
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "user"
        required: false
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
...
  /user/v1/{id}:
    get:
      tags:
      - "User"
      summary: "Get a user by ID."
      description: ""
      operationId: "getUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
    put:
      tags:
      - "User"
      summary: "Update a user."
      description: ""
      operationId: "updateUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "user"
        required: false
        schema:
          $ref: "#/definitions/User"
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
    delete:
      tags:
      - "User"
      summary: "Delete a user by ID."
      description: ""
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
definitions:
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      first_name:
        type: "string"
      last_name:
        type: "string"
      comments:
        type: "string"
      email:
        type: "string"
      state:
        type: "string"
      password:
        type: "string"
      password_hash:
        type: "string"
      login_failures:
        type: "integer"
        format: "int32"
      login_count:
        type: "integer"
        format: "int32"
      last_login:
        type: "string"
      updated_on:
        type: "string"
      updated_by_user:
        type: "string"
      created_on:
        type: "string"
      created_by_user:
        type: "string"